// ProjetoPokedex/projeto-pokedex/src/types/PokemonCardTypes.ts
interface PokemonData {
  id: number
  name: string
  amount: number
}

interface Pokemon {
  name: string
  url: string
}

interface PokemonCardContextData {
  pokemonCart: PokemonData[]
  pokemonAmount: number
  searchByNamePokemon: (name: string) => void
  addToPokemon: (pokemon: Pokemon, name: string) => void
  removePokemon: (id: number) => void
  removeAllClearPokemon: () => void
  pokemonsPerPage: number
  getCurrentPokemons: () => PokemonData[]
  nextPage: () => void
  prevPage: () => void
}

export { PokemonData, Pokemon, PokemonCardContextData }